app/processor/requirements.txt
weaviate-client==3.24.1
PyPDF2==3.0.1
tqdm==4.65.0
langchain==0.0.335
python-dotenv==1.0.0
